test: combine filter and filter not tests into 1 file
