+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/testutils/commonFilterToSetTests.kt
@@ -1,5 +1,6 @@
 package xyz.lbres.kotlinutils.set.multiset.testutils
 
+import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSet
 import xyz.lbres.kotlinutils.set.multiset.emptyMultiSet
@@ -14,6 +15,18 @@ private val e2 = ArithmeticException()
 private val e3 = ClassCastException("Cannot cast Int to List")
 private val e4 = ClassCastException("other message")
 
+private var previousOdd = false
+private val intPredicate: (Int) -> Boolean = {
+    when {
+        it % 2 == 0 -> true
+        previousOdd -> false
+        else -> {
+            previousOdd = true
+            true
+        }
+    }
+}
+
 private fun <S> runSingleTest(set: MultiSet<S>, expected: MultiSet<S>, const: Boolean, genericFilter: GenericFilterFn<*>, filterFn: (S) -> Boolean) {
     val result = genericFilter(set, filterFn)
     assertEquals(expected, result)
@@ -26,17 +39,7 @@ fun runCommonFilterToSetTests(createSet: (Collection<*>) -> MultiSet<*>, generic
 
     val intSet = createIntSet(listOf(1, 1, 2, 14, 14))
     val intExpected = multiSetOf(1, 2, 14, 14)
-    var previousOdd = false
-    val intPredicate: (Int) -> Boolean = {
-        when {
-            it % 2 == 0 -> true
-            previousOdd -> false
-            else -> {
-                previousOdd = true
-                true
-            }
-        }
-    }
+    previousOdd = false
     val result = genericFilter(intSet, intPredicate)
     assertEquals(intExpected, result)
 }
@@ -47,20 +50,49 @@ fun runCommonFilterToSetConsistentTests(createSet: (Collection<*>) -> MultiSet<*
 
     val intSet = createIntSet(listOf(1, 1, 3, 2, 14, 14))
     val intOptions = listOf(multiSetOf(1, 1, 2, 14, 14), multiSetOf(3, 2, 14, 14))
-    var previousOdd = false
-    val intPredicate: (Int) -> Boolean = {
-        when {
-            it % 2 == 0 -> true
-            previousOdd -> false
-            else -> {
-                previousOdd = true
-                true
-            }
-        }
-    }
+    previousOdd = false
     val result = genericFilter(intSet, intPredicate)
     assertContains(intOptions, result)
     assertEquals(const, result is ConstMultiSet<*>)
+}
+
+fun runCommonFilterNotToSetTests(createSet: (Collection<*>) -> MultiSet<*>, genericFilter: GenericFilterFn<*>, const: Boolean) {
+    runCommonNotTests(createSet, genericFilter, const)
+    val createIntSet = getCreateSet<Int>(createSet)
+
+    val intSet = createIntSet(listOf(1, 1, 3, 2, 14, 14))
+    val intOptions = listOf(multiSetOf(1, 1), multiSetOf(1, 3))
+    previousOdd = false
+    val result = genericFilter(intSet, intPredicate)
+    assertContains(intOptions, result)
+    assertEquals(const, result is ConstMultiSet<*>)
+}
+
+fun runCommonFilterNotToSetConsistentTests(createSet: (Collection<*>) -> MultiSet<*>, genericFilter: GenericFilterFn<*>, const: Boolean) {
+    runCommonNotTests(createSet, genericFilter, const)
+    val createIntSet = getCreateSet<Int>(createSet)
+    val createIntListSet = getCreateSet<IntList>(createSet)
+
+    val intSet = createIntSet(listOf(1, 1, 3, 2, 14, 14))
+    val intOptions = listOf(multiSetOf(1, 1), multiSetOf(3))
+    previousOdd = false
+    val intResult = genericFilter(intSet, intPredicate)
+    assertContains(intOptions, intResult)
+    assertEquals(const, intResult is ConstMultiSet<*>)
+
+    val listSet = createIntListSet(listOf(listOf(1, 2, 3), listOf(0, 5, 7), listOf(1, 2, 3)))
+
+    var previous12 = false
+    val listPredicate: (IntList) -> Boolean = {
+        if (it.containsAll(listOf(1, 2)) && previous12) {
+            false
+        } else {
+            previous12 = true
+            true
+        }
+    }
+    val listExpected: MultiSet<IntList> = emptyMultiSet()
+    runSingleTest(listSet, listExpected, const, genericFilter, listPredicate)
 }
 
 private fun runCommonTests(createSet: (Collection<*>) -> MultiSet<*>, genericFilter: GenericFilterFn<*>, const: Boolean) {
@@ -93,4 +125,41 @@ private fun runCommonTests(createSet: (Collection<*>) -> MultiSet<*>, genericFil
     val errorSet: MultiSet<Exception> = createExceptionSet(listOf(e1, e1, e2, e3, e4, e4))
     val errorExpected: MultiSet<Exception> = multiSetOf(e3, e4, e4)
     runSingleTest(errorSet, errorExpected, const, genericFilter) { it is ClassCastException }
+}
+
+private fun runCommonNotTests(createSet: (Collection<*>) -> MultiSet<*>, genericFilter: GenericFilterFn<*>, const: Boolean) {
+    val createIntSet = getCreateSet<Int>(createSet)
+    val createIntListSet = getCreateSet<IntList>(createSet)
+    val createStringSet = getCreateSet<String>(createSet)
+    val createExceptionSet = getCreateSet<Exception>(createSet)
+
+    var intSet = createIntSet(emptyList())
+    var intExpected = emptyMultiSet<Int>()
+    runSingleTest(intSet, intExpected, const, genericFilter) { true }
+
+    intSet = createIntSet(listOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0))
+    intExpected = emptyMultiSet()
+    runSingleTest(intSet, intExpected, const, genericFilter) { true }
+
+    intExpected = createIntSet(listOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0))
+    runSingleTest(intSet, intExpected, const, genericFilter) { false }
+
+    intExpected = createIntSet(listOf(2, 3, 4, 6, -1))
+    runSingleTest(intSet, intExpected, const, genericFilter) { intSet.getCountOf(it) > 1 }
+
+    intExpected = createIntSet(listOf(2, 4, 5, 5, 0, 0))
+    val testInts = setOf(-1, 1, 11, 12, 10, 3, 6, -2)
+    runSingleTest(intSet, intExpected, const, genericFilter) { it in testInts }
+
+    val stringSet = createStringSet(listOf("abc", "abc", "hello", "world", "goodbye", "world", "hi", "world"))
+    val stringExpected = multiSetOf("goodbye", "hi")
+    runSingleTest(stringSet, stringExpected, const, genericFilter) { it.length in 3..5 }
+
+    val errorSet = createExceptionSet(listOf(e1, e1, e2, e3, e4, e4))
+    val errorExpected: MultiSet<Exception> = multiSetOf(e1, e1, e2)
+    runSingleTest(errorSet, errorExpected, const, genericFilter) { it is ClassCastException }
+
+    val listSet = createIntListSet(listOf(listOf(1, 2, 3), listOf(0, 5, 6)))
+    val listExpected = multiSetOf(listOf(1, 2, 3))
+    runSingleTest(listSet, listExpected, const, genericFilter) { it.contains(0) }
 }
